requirements.txt
langchain-core==0.3.0
langchain-community==0.3.0
langchain-openai==0.2.0
langchain-anthropic==0.2.0
langgraph==0.2.22
pydantic==2.9.2
pydantic-settings==2.5.2
retry==0.9.2
numpy==1.26.4
faiss-cpu==1.8.0.post1
tavily-python==0.5.0
wikipedia==1.4.0
google-genai

python-dotenv==1.0.1
